+++ CCD_prototype/CCD/makeintmat.h
@@ -214,8 +214,8 @@ public:
     int                             Identity_hhhp(int h1, int h2, int h3, int p1);
     int                             Identity_hhpp(int h1, int h2, int p1, int p2);
     int                             Identity_ppph(int p1, int p2, int p3, int h1);
-    unsigned long int          Identity_hhhppp(int h1, int h2, int h3, int p1, int p2, int p3);
-    int                             Identity_hhhhhp(int h1, int h2, int h3, int h4, int h5, int p1); //this is a special function for T3e alone
+    unsigned long int               Identity_hhhppp(int h1, int h2, int h3, int p1, int p2, int p3);
+    unsigned long int               Identity_hhhhhp(int h1, int h2, int h3, int h4, int h5, int p1); //this is a special function for T3e alone
     spp::sparse_hash_map<int, double>           Vhhhp_elements; //needed for T3
     spp::sparse_hash_map<int, double>           Vhhpp_elements;
     spp::sparse_hash_map<int, double>           Vppph_elements; //needed for T3
